Add validation for AKO ipfamily and Cluster node ipfamily
